Split d3d12 functions in their own files
