+++ runtime_openxr/src/d3d11compositor.cpp
@@ -5,6 +5,7 @@
 #include "filesystem"
 #include "settings.h"
 #include "d3d11swapchain.h"
+#include "swapchain.h"
 
 namespace fs = std::filesystem;
