+++ runtime_openxr/src/d3d11swapchain.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "swapchain.h"
 #include "xrrendering.h"
 #include "openxr_includes.h"
